SCRUM-3121 moved additional field data components
