+++ src/main/cliapp/src/components/FieldData/AssertedAlleleAdditionalFieldData.js
