+++ src/main/cliapp/src/components/FieldData/DiseaseAdditionalFieldData.js
