+++ src/main/cliapp/src/components/FieldData/SGDStrainBackgroundAdditionalFieldData.js
