+++ src/main/cliapp/src/components/FieldData/TaxonAdditionalFieldData.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+export const TaxonAdditionalFieldData = ({ curie }) => {
+  if(curie){
+    return <div className="p-info">{curie}</div>;
+  } else {
+    return null;
+  } 
+}
